+++ app/Models/Policies.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class Policies extends Model
+{
+    protected $guarded = [];
+
+    public function categories() {
+        return $this->belongsTo(PolicyCategories::class, 'policy_category_id'. 'id');
+    }
+}