+++ database/migrations/2026_05_18_011551_create_policies_table.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('policies', function (Blueprint $table) {
+            $table->id();
+            $table->unsignedBigInteger('policy_category_id');
+            $table->foreign('policy_category_id')->references('id')->on('policy_categories')->onDelete('restrict')->onUpdate('cascade');
+            $table->string('title');
+            $table->text('body');
+            $table->boolean('is_active')->default(true);
+            $table->softDeletes();
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('policies');
+    }
+};