+++ resources/views/master/librarian_master.blade.php
@@ -83,6 +83,102 @@
                         <span class="nav-link-title"> Dashboard </span>
                       </a>
                     </li>
+                    <li class="nav-item">
+                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-book-2">
+                          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                          <path d="M19 4v16h-12a2 2 0 0 1 -2 -2v-12a2 2 0 0 1 2 -2h12" />
+                          <path d="M19 16h-12a2 2 0 0 0 -2 2" />
+                          <path d="M9 8h6" />
+                        </svg>
+                        <span class="nav-link-title"> Borrows </span>
+                      </a>
+                    </li>
+                    <li class="nav-item">
+                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                        <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/home -->
+                            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-api-book">
+                              <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                              <path d="M3 19a9 9 0 0 1 9 0a9 9 0 0 1 1.006 -.5" />
+                              <path d="M3 6a9 9 0 0 1 9 0a9 9 0 0 1 9 0" />
+                              <path d="M3 6v13" />
+                              <path d="M12 6v13" />
+                              <path d="M21 6v6" />
+                              <path d="M17.001 19a2 2 0 1 0 4 0a2 2 0 1 0 -4 0" />
+                              <path d="M19.001 15.5v1.5" />
+                              <path d="M19.001 21v1.5" />
+                              <path d="M22.032 17.25l-1.299 .75" />
+                              <path d="M17.27 20l-1.3 .75" />
+                              <path d="M15.97 17.25l1.3 .75" />
+                              <path d="M20.733 20l1.3 .75" />
+                            </svg>
+                        </span>
+                        <span class="nav-link-title"> Manage Books </span>
+                      </a>
+                    </li>
+                    <li class="nav-item">
+                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                        <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/home -->
+                            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-user-edit">
+                              <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                              <path d="M8 7a4 4 0 1 0 8 0a4 4 0 0 0 -8 0" />
+                              <path d="M6 21v-2a4 4 0 0 1 4 -4h3.5" />
+                              <path d="M18.42 15.61a2.1 2.1 0 0 1 2.97 2.97l-3.39 3.42h-3v-3l3.42 -3.39" />
+                            </svg>
+                        </span>
+                        <span class="nav-link-title"> Authors </span>
+                      </a>
+                    </li>
+                    <li class="nav-item">
+                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                        <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/home -->
+                            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-user-cog">
+                              <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                              <path d="M8 7a4 4 0 1 0 8 0a4 4 0 0 0 -8 0" />
+                              <path d="M6 21v-2a4 4 0 0 1 4 -4h2.5" />
+                              <path d="M17.001 19a2 2 0 1 0 4 0a2 2 0 1 0 -4 0" />
+                              <path d="M19.001 15.5v1.5" />
+                              <path d="M19.001 21v1.5" />
+                              <path d="M22.032 17.25l-1.299 .75" />
+                              <path d="M17.27 20l-1.3 .75" />
+                              <path d="M15.97 17.25l1.3 .75" />
+                              <path d="M20.733 20l1.3 .75" />
+                            </svg>
+                        </span>
+                        <span class="nav-link-title"> Patrons </span>
+                      </a>
+                    </li>
+                    <li class="nav-item">
+                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                        <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/home -->
+                            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-user-exclamation">
+                              <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                              <path d="M8 7a4 4 0 1 0 8 0a4 4 0 0 0 -8 0" />
+                              <path d="M6 21v-2a4 4 0 0 1 4 -4h4c.348 0 .686 .045 1.008 .128" />
+                              <path d="M19 16v3" />
+                              <path d="M19 22v.01" />
+                            </svg>
+                        </span>
+                        <span class="nav-link-title"> Patron Types </span>
+                      </a>
+                    </li>
+                    <li class="nav-item">
+                      <a class="nav-link" href="{{ route('librarian.dashboard') }}">
+                        <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/home -->
+                          <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-report">
+                            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+                            <path d="M8 5h-2a2 2 0 0 0 -2 2v12a2 2 0 0 0 2 2h5.697" />
+                            <path d="M18 14v4h4" />
+                            <path d="M18 11v-4a2 2 0 0 0 -2 -2h-2" />
+                            <path d="M8 5a2 2 0 0 1 2 -2h2a2 2 0 0 1 2 2a2 2 0 0 1 -2 2h-2a2 2 0 0 1 -2 -2" />
+                            <path d="M14 18a4 4 0 1 0 8 0a4 4 0 1 0 -8 0" />
+                            <path d="M8 11h4" />
+                            <path d="M8 15h3" />
+                          </svg>
+                        </span>
+                        <span class="nav-link-title"> Reports </span>
+                      </a>
+                    </li>
                   </ul>
                   <!-- END NAVBAR MENU -->
                 </div>